+++ src/main/java/lotto/Rank.java
@@ -1,0 +1,5 @@
+package lotto;
+
+public class Rank {
+
+}
